+++ config/wallet-config.json
@@ -1,6 +1,6 @@
 {
   "activeFiatProviders": {
     "transak": { "name": "transak", "enabled": true },
-    "okcoin": { "name": "okcoin", "enabled": false }
+    "okcoin": { "name": "okcoin", "enabled": true }
   }
 }
