+++ src/app/features/fiat-onramp-providers/okcoin-helper.ts
@@ -1,7 +1,5 @@
-const isProduction = process.env.WALLET_ENVIRONMENT === 'production';
-
-const subdomain = isProduction ? 'www' : 'beta';
-
 export function makeOkcoinUrl(address: string) {
-  return `https://${subdomain}.okcoin.org/bridgeBeta/thirdBridge?isThirdBridge=1&partnerId=10002&crypto=STX&address=w${address}&thirdPartyName=Hiro&fiatAmount=100&currency=USD`;
+  const successBackLink = encodeURI(`https://explorer.stacks.co/address/${address}?chain=mainnet`);
+  const thirdPartyLink = encodeURI('https://hiro.so');
+  return `https://www.okcoin.com/bridge/thirdBridge?isThirdBridge=1&partnerId=10002&crypto=STX&address=${address}&thirdPartyName=Hiro&fiatAmount=100&currency=USD&successBackLink=${successBackLink}&thirdPartyLink=${thirdPartyLink}`;
 }
